requirements.txt
yfinance==0.1.70
pandas==1.3.3
matplotlib==3.4.3
PyYAML==5.4.1